requirements.txt
wheel
pyyaml 
psutil
paramiko
fastapi
uvicorn[standard]